apps/exploratory-copilot-app/backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
aiofiles
python-dotenv
langchain
langchain-openai
langgraph
openai
pandas
numpy
plotly
matplotlib
missingno
sweetviz
dtale
sqlalchemy
ai-data-science-team
